requirements.txt
requests==2.28.1
pandas==1.5.2
kafka-python==2.0.2
pymysql==1.0.2
pyspark==3.5.1